pyproject.toml
[tool.poetry]
name = "env-parser"
version = "0.1.0"
description = "A friendly enviroment variable parser to describe envs and parse them. This will make documenting environment variables easier."
authors = ["Victoria Luquet <victoria.tewari@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rich = "^13.7.1"
pydantic = "^2.7.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
